fix: исправления по замечаниям ревьюера (#3)
