+++ src/test/java/ru/t1/ismailov/taskmanager/kafka/EnableKafkaTestContainer.java
@@ -1,5 +1,8 @@
 package ru.t1.ismailov.taskmanager.kafka;
 
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 
 import java.lang.annotation.ElementType;
@@ -9,6 +12,9 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
+@ActiveProfiles("test")
+@AutoConfigureMockMvc
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @ContextConfiguration(initializers = KafkaTestContainer.Initializer.class)
 public @interface EnableKafkaTestContainer {
 }
