+++ src/test/java/ru/t1/ismailov/taskmanager/repository/TaskRepositoryTest.java
@@ -4,15 +4,13 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import ru.t1.ismailov.taskmanager.model.Task;
 import ru.t1.ismailov.taskmanager.model.TaskStatus;
 
-@DataJpaTest
 @ActiveProfiles("test")
-@Testcontainers
+@SpringBootTest
 class TaskRepositoryTest {
 
     @Autowired
