+++ src/test/resources/fill_users_for_update_test.sql
@@ -4,4 +4,5 @@ TRUNCATE TABLE tasks RESTART IDENTITY CASCADE;
 -- 2. Наполняем таблицу тестовыми задачами
 INSERT INTO tasks (title, description, user_id, status)
 VALUES ('title_1', 'desc_1', 1, 'NEW'),
-       ('title_2', 'desc_1', 2, 'UPDATING');
+       ('title_2', 'desc_2', 2, 'UPDATING'),
+       ('title_3', 'desc_3', 3, 'NEW');
